client tab created; partial class for main windows introduced
